enforce dtypes

+++ numpywren/matrix.py
@@ -258,7 +258,7 @@ class BigMatrix(object):
             X_block = self.parent_fn(self, *block_idx)
         else:
             bio = self.__s3_key_to_byte_io__(key)
-            X_block = np.load(bio)
+            X_block = np.load(bio).astype(self.dtype)
         if (self.transposed):
             X_block = X_block.T
         return X_block
@@ -271,6 +271,8 @@ class BigMatrix(object):
             raise Exception("Incompatible block size: {0} vs {1}".format(block.shape, current_shape))
         if (self.transposed):
             block = block.T
+
+        block = block.astype(self.dtype)
         key = self.__shard_idx_to_key__(block_idx)
         return self.__save_matrix_to_s3__(block, key)
 
@@ -386,7 +388,7 @@ class BigSymmetricMatrix(BigMatrix):
             X_block = self.parent_fn(self, *block_idx_sym)
         else:
             bio = self.__s3_key_to_byte_io__(key)
-            X_block = np.load(bio)
+            X_block = np.load(bio).astype(self.dtype)
         if (flipped):
             X_block = X_block.T
         return X_block
@@ -401,6 +403,7 @@ class BigSymmetricMatrix(BigMatrix):
         if (block.shape != current_shape):
             raise Exception("Incompatible block size: {0} vs {1}".format(block.shape, current_shape))
         key = self.__shard_idx_to_key__(block_idx)
+        block = block.astype(self.dtype)
         return self.__save_matrix_to_s3__(block, key)
